Fixed gradient polyline not always fully drawn + crashes (#1960)
The gradient polyline was not always completely visible, because the Matrix transform in the paths was not applied correctly. This happened, because the segments were cached and could have the wrong matrix transform applied to them. This could cause the line to not visible, because it will below the map-tile.
Also, I’ve witnessed some rare crashes due to the caching of the segments. Apparently, the class was accessed from multiple threads which led to threading issues. Removing the segments cache fixed this problem as well.